Move credential-specific functionality to own package
