+++ pkg/secrethub/account_key.go
@@ -3,6 +3,7 @@ package secrethub
 import (
 	"github.com/secrethub/secrethub-go/internals/api"
 	"github.com/secrethub/secrethub-go/internals/crypto"
+	"github.com/secrethub/secrethub-go/pkg/secrethub/credentials"
 )
 
 // DefaultAccountKeyLength defines the default bit size for account keys.
@@ -15,7 +16,7 @@ func generateAccountKey() (crypto.RSAPrivateKey, error) {
 // AccountKeyService handles operations on SecretHub account keys.
 type AccountKeyService interface {
 	// Create creates an account key for the client's credential.
-	Create(credentialFingerprint string, encrypter Encrypter) (*api.EncryptedAccountKey, error)
+	Create(credentialFingerprint string, encrypter credentials.Encrypter) (*api.EncryptedAccountKey, error)
 	// Exists returns whether an account key exists for the client's credential.
 	Exists() (bool, error)
 }
@@ -32,7 +33,7 @@ func newAccountKeyService(client *Client) accountKeyService {
 }
 
 // Create creates an account key for the clients credential.
-func (s accountKeyService) Create(credentialFingerprint string, encrypter Encrypter) (*api.EncryptedAccountKey, error) {
+func (s accountKeyService) Create(credentialFingerprint string, encrypter credentials.Encrypter) (*api.EncryptedAccountKey, error) {
 	key, err := generateAccountKey()
 	if err != nil {
 		return nil, err
