+++ pkg/secrethub/user.go
@@ -4,6 +4,7 @@ import (
 	"github.com/secrethub/secrethub-go/internals/api"
 	"github.com/secrethub/secrethub-go/internals/crypto"
 	"github.com/secrethub/secrethub-go/internals/errio"
+	"github.com/secrethub/secrethub-go/pkg/secrethub/credentials"
 )
 
 // UserService handles operations on users from SecretHub.
@@ -11,7 +12,7 @@ type UserService interface {
 	// Me gets the account's user if it exists.
 	Me() (*api.User, error)
 	// Create creates a new user at SecretHub.
-	Create(username, email, fullName string, verifier Verifier, encrypter Encrypter) (*api.User, error)
+	Create(username, email, fullName string, verifier credentials.Verifier, encrypter credentials.Encrypter) (*api.User, error)
 	// Get a user by their username.
 	Get(username string) (*api.User, error)
 }
@@ -32,7 +33,7 @@ func (s userService) Me() (*api.User, error) {
 }
 
 // Create creates a new user at SecretHub.
-func (s userService) Create(username, email, fullName string, verifier Verifier, encrypter Encrypter) (*api.User, error) {
+func (s userService) Create(username, email, fullName string, verifier credentials.Verifier, encrypter credentials.Encrypter) (*api.User, error) {
 	err := api.ValidateUsername(username)
 	if err != nil {
 		return nil, errio.Error(err)
@@ -56,7 +57,7 @@ func (s userService) Create(username, email, fullName string, verifier Verifier,
 	return s.create(username, email, fullName, accountKey, verifier, encrypter)
 }
 
-func (s userService) create(username, email, fullName string, accountKey crypto.RSAPrivateKey, verifier Verifier, encrypter Encrypter) (*api.User, error) {
+func (s userService) create(username, email, fullName string, accountKey crypto.RSAPrivateKey, verifier credentials.Verifier, encrypter credentials.Encrypter) (*api.User, error) {
 	credentialRequest, err := s.client.createCredentialRequest(verifier)
 	if err != nil {
 		return nil, errio.Error(err)
@@ -105,7 +106,7 @@ func (s userService) Get(username string) (*api.User, error) {
 }
 
 // createAccountKey adds the account key for the clients credential.
-func (c *Client) createAccountKey(credentialFingerprint string, accountKey crypto.RSAPrivateKey, encrypter Encrypter) (*api.EncryptedAccountKey, error) {
+func (c *Client) createAccountKey(credentialFingerprint string, accountKey crypto.RSAPrivateKey, encrypter credentials.Encrypter) (*api.EncryptedAccountKey, error) {
 	accountKeyRequest, err := c.createAccountKeyRequest(encrypter, accountKey)
 	if err != nil {
 		return nil, errio.Error(err)
